requirements.txt
Django
mysqlclient
gunicorn
django-cors-headers 
django-two-factor-auth
phonenumbers